User Mode Linux support
Add support for building a User Mode Linux kernel (executable) that
can be used together with the rootfs to run a virtual instance as a
regular process, no need for QEMU or such things. This can be very
useful for testing software that require root capabilities on the
kernel, i.e. kernel modules or applications using eBPF.